Add files via upload
if there is an "i." before the name of a folder, it just means that those projects were from the intermediate lessons of whatever language it contains. also, if there are some projects that skipped numbers ( i skipped i.css 2 and i.css 3) that means i either skipped the project for some reason or something went wrong with the files.